question4-5/requirements.txt
Twisted>=21.2.0
autobahn>=21.2.1
pandas
pyOpenSSL>=19.0.0
requests>=2.28.1
requests>=2.25.1
service-identity>=21.1.0
simplejson>=3.17.6
binance-connector>=1.14.0
